Fix persistent enable flag in the cloud when error occur.
When the HLS and WebRTC are enabled at the same time and after that immediately one of them got disabled then the error message will appear but the enable flag in the cloud will still persist in enable state.

It was fixed by sending the disable flag to the cloud when the error is detected.